Commit the first version of the Gareth Vagrant setup.
This is functionally able to `vagrant up` and install enough for Gareth to start running on http://localhost:8080/ with an admin:vagrant user.
We still need a README. There are various @fixmes and @todos for things in the puppet config that could be cleaned up and done better.
We also need to setup puppet config for a STOMP server. Right now Gareth works and you can create repos. But because of the missing STOMP
server trying to fetch a repo will error out instead of working properly.
Likewise the task runner is not being executed.

Ori Livneh's work on mediawiki-vagrant was used as a reference when putting together various parts of this Vagrant setup.